% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_year_frac}
\alias{add_year_frac}
\title{Add 'year fraction' (yrFrac) to a data table}
\usage{
add_year_frac(dt)
}
\arguments{
\item{dt}{a data table to which yrFrac will be added}
}
\value{
the data table with yrFrac added
}
\description{
So-called 'year fraction' is the percentage of hours that have elapsed
given the date and hour. For example at 2 am on 3/1/17, yrFrac = 0.16187:
(59 days * 24 hours + 2 hours)/8760 hours in a year.
}
