% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avgPinball}
\alias{avgPinball}
\title{Calculate the average pinball across multiple quantiles}
\usage{
avgPinball(qdt, quants)
}
\arguments{
\item{qdt}{a data table containing actuals and predicted quantiles}

\item{quants}{the target quantiles \eqn{0 < quantile < 1} to assess}
}
\value{
the average pinball score across the samples and quantiles
}
\description{
\code{avgPinball} calls \code{pinball} across a range of
quantiles and returns the average
}
\examples{
\dontrun{
qdt = data.table(act=rep(2700,3),
                 q0.1 = c(2600,2500,2500),
                 q0.5=c(2900,2800,2700),
                 q0.9 = c(3100, 3000, 2900))
qnts = c(0.1, 0.5, 0.9)
avgPinball( qdt, qnts) # should return 32.22
}
}
