% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_isone_data.R
\name{clean_isone_data}
\alias{clean_isone_data}
\title{Clean the ISONE data}
\usage{
clean_isone_data(cln_yaml_cfg = "")
}
\arguments{
\item{cln_yaml_cfg}{a yaml file with configuration settings}
}
\description{
\code{clean_isone_data} removes variables that aren't needed, then adds some
derived weather variables, and potentially useful interactions. It then
writes an RDS file named and to a directory specified in the configuration
file.
}
\examples{
\dontrun{
clean_isone_data('/home/wally/gefcom17/config/gen_and_dates_cfg.yaml')
}
}
