% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_isone_data.R
\name{combine_isone_data}
\alias{combine_isone_data}
\title{Combine the xls(x)'s downloaded from ISONE's website}
\usage{
combine_isone_data(yaml_cfg = "")
}
\arguments{
\item{yaml_cfg}{a yaml file with configuration settings}
}
\description{
\code{combine_isone_data} combines the yearly excel file retrieved
from ISONE's website into one RDS file. In the process it standardizes
column names, and deals with DST start/end.
}
\details{
\strong{Note} the output file for this function is the input file
for \code{clean_isone_data}
}
\examples{
\dontrun{
combine_isone_data('/home/wally/gefcom17/config/gen_and_dates_cfg.yaml')
}
}
