% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fact_to_int}
\alias{fact_to_int}
\title{Convert factor variables into integers}
\usage{
fact_to_int(dt)
}
\arguments{
\item{dt}{a data table possibly containing factors variables}
}
\value{
the data table with any factors converted in integers
}
\description{
Because xgboost, at least the way I used it, does not work with non-numerical
data \code{fact_to_int} converts factors to integers.  This function was
more important when I thought inclusion of interactions, e.g.
hour x weekend/weekday, would help performance.  It turned out not to be the
case.
}
