% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_isone_smd_data}
\alias{get_isone_smd_data}
\title{Get SMD data from ISO New England (ISONE)'s website}
\usage{
get_isone_smd_data(minYr = 2016, maxYr = year(Sys.Date()), data_dir = "")
}
\arguments{
\item{minYr}{an integer, the first year for which you want data, defaults to 2011}

\item{maxYr}{an integer, the last year for which you want data, defaults to current year}

\item{data_dir}{a string, the directory in which to write files,
without the trailing '/'}
}
\description{
Note: due to lack of file-naming consistency, this function only
works for years 2016 and later.  See the isone website referenced
in the code for more information.
}
\details{
\code{get_isone_smd_data} retrieves hourly load and price data from
ISONE's website, and writes the yearly files to the directory specified
}
\examples{
\dontrun{
get_isone_smd_data(2016, 2018, '/home/frodo/isone/smd_data')
}
}
