% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wx.R
\name{model_temperature}
\alias{model_temperature}
\title{Produce estimates of quantiles of temperature}
\usage{
model_temperature(wx_cfg = "", dts_cfg = "")
}
\arguments{
\item{wx_cfg}{a yaml file containing modeling parameters}

\item{dts_cfg}{a yaml file containing general setup parameters}
}
\description{
\code{model_temperature} produces quantile estimates of
temperature.  This function is driven by yaml configuration files,
examples of which are included with this package.  The configuration
files also explain the meaning/import/use of the parameters.
}
\details{
\strong{Note:} The gamboost specifications are largely hard-coded.  I could
not figure out how to parameterize the function call with respect to
the formula: gamboost( tc ~ bbs(...) + ...)

If writeFiles = 1 (in the modeling configuration file), then predictions
and results are written to the directories specified in the configuration
file.
}
\examples{
\dontrun{
model_weather('~/gefcom17/config/wx_cfg.yaml',
'~/gefcom17/config/gen_and_dates_cfg.yaml')
}
}
