% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pinball}
\alias{pinball}
\title{Calculate the pinball loss between actual and the predicted quantile}
\usage{
pinball(y_act, y_pred, qu, dosum = 1)
}
\arguments{
\item{y_act}{a single number or numerical vector of actual quantities}

\item{y_pred}{the predicted quantile of that quantity}

\item{qu}{the target quantile \eqn{0 < qu < 1}}

\item{dosum}{when 1, return the pinball loss as a single number}
}
\value{
a vector of pinball values, or their sum, depending on value of dosum
}
\description{
\code{pinball} calculates the goodness of fit between an actual quantity,
e.g. electricity demand, and a predicted quantile of that quantity.
}
\examples{
\dontrun{
act = rep(2700, 3)
pred = rep(2671, 3)
ploss = pinball( act, pred, 0.1)
# should return 8.7
}
}
